Move camera stuff out of the game module
