+++ src/camera/mod.rs
@@ -1,20 +1,33 @@
+//! Main rendering camera, its movement, and related effects
+
+mod inputs;
+pub mod parallax;
+
+use crate::{
+	game::{prelude::*, spawn::SpawnLevel},
+	graphics::*,
+	screen::{DoScreenTransition, Screen},
+	AppSet,
+};
 use bevy::{
 	math::bounding::{Aabb2d, BoundingVolume},
 	render::camera::ScalingMode,
 };
+use inputs::{MoveCameraEvent, ZoomCameraEvent};
 
-use crate::{
-	game::prelude::*,
-	graphics::{
-		GAME_AREA, LEVEL_AREA_CENTER, LEVEL_AREA_WIDTH, SPRITE_SIZE, VERTICAL_PADDING_FRACTION,
-	},
-	screen::{DoScreenTransition, Screen},
-	AppSet,
-};
-
-use crate::game::spawn::SpawnLevel;
-
-use super::inputs::{MoveCameraEvent, ZoomCameraEvent};
+pub(super) fn plugin(app: &mut App) {
+	app.add_plugins((inputs::plugin, parallax::plugin))
+		.add_systems(Startup, spawn_camera)
+		.add_systems(
+			Update,
+			(
+				set_camera_on_screen_transition.run_if(on_event::<DoScreenTransition>),
+				set_camera_level_view.run_if(on_event::<SpawnLevel>),
+				update_camera.after(AppSet::RecordInput),
+			)
+				.chain(),
+		);
+}
 
 #[derive(Component, Clone, Copy)]
 #[require(Camera, Camera2d, CameraIntertia)]
@@ -25,40 +38,6 @@ pub struct CameraHarness {
 	pub level_bounds: Aabb2d,
 	/// The world-coordinate point the camera is looking at in the center of its view.
 	pub center: Vec2,
-}
-
-/// Allows turning off the parallax effect, because this seems exactly like
-/// something that we want to be configurable
-#[derive(Resource, Clone, Copy, PartialEq, Eq)]
-pub struct EnableParallax(pub bool);
-
-impl Default for EnableParallax {
-	fn default() -> Self {
-		// Parallax is enabled by default
-		Self(true)
-	}
-}
-
-/// This component applies a parallax effect to the [`Transform`] of an entity.
-/// The value of parallax should be in [0, 1] (0 means no parallax,
-/// 1 means the entity will follow the camera).
-///
-/// The parallaxed entity's [`Transform`] must not be changed
-/// by other systems. Parallax only works on static entities.
-///
-/// Assumes there is only one camera. We can worry about the alternative
-/// when we need another camera.
-///
-/// This cannot be applied to a camera entity, for obvious reasons.
-#[derive(Component, Clone, Copy)]
-pub struct Parallax(pub f32);
-
-/// Support component for [`Parallax`], records the default position
-/// and scale of the entity before parallax was applied.
-#[derive(Component, Clone, Copy)]
-struct ParallaxBasis {
-	position: Vec2,
-	scale: Vec2,
 }
 
 /// Support component for [`CameraHarness`] that tracks
@@ -78,26 +57,6 @@ impl Default for CameraHarness {
 			center: LEVEL_AREA_CENTER,
 		}
 	}
-}
-
-pub fn plugin(app: &mut App) {
-	app
-		// Spawn the main camera.
-		.add_systems(Startup, spawn_camera)
-		.add_systems(
-			Update,
-			(
-				init_parallax,
-				(
-					set_camera_on_screen_transition.run_if(on_event::<DoScreenTransition>),
-					set_camera_level_view.run_if(on_event::<SpawnLevel>),
-					update_camera.after(AppSet::RecordInput),
-					apply_paralax.run_if(resource_equals(EnableParallax(true))),
-				)
-					.chain(),
-			),
-		)
-		.init_resource::<EnableParallax>();
 }
 
 fn spawn_camera(mut commands: Commands) {
@@ -289,30 +248,4 @@ fn update_camera(
 
 	transform.translation.x = harness.center.x;
 	transform.translation.y = harness.center.y;
-}
-
-fn init_parallax(mut commands: Commands, query: Query<(Entity, &Transform), Added<Parallax>>) {
-	for (id, transform) in &query {
-		commands.entity(id).insert(ParallaxBasis {
-			position: transform.translation.xy(),
-			scale: transform.scale.xy(),
-		});
-	}
-}
-
-fn apply_paralax(
-	mut query: Query<(&mut Transform, &Parallax, &ParallaxBasis), Without<Camera2d>>,
-	camera_q: Query<(&Transform, &CameraHarness), (With<Camera2d>, Changed<Transform>)>,
-) {
-	let Ok((camera_transform, camera_harness)) = camera_q.single() else {
-		return;
-	};
-	for (mut transform, Parallax(parallax), basis) in &mut query {
-		let new_scale = basis.scale / camera_harness.scale.powf(*parallax);
-		let new_position = basis.position + camera_transform.translation.xy() * parallax;
-		transform.translation.x = new_position.x;
-		transform.translation.y = new_position.y;
-		transform.scale.x = new_scale.x;
-		transform.scale.y = new_scale.y;
-	}
 }
