+++ src/game/mod.rs
@@ -1,7 +1,6 @@
 //! Game mechanics and content.
 
 mod animation;
-pub mod camera;
 pub mod components;
 mod drawing;
 mod game_sfx;
@@ -41,6 +40,5 @@ pub(super) fn plugin(app: &mut App) {
 		game_sfx::plugin,
 		history::plugin,
 		hints::plugin,
-		camera::plugin,
 	));
 }
